Add files via upload
This is a code that calculates the curvature, or stacks of images describing t-cells. The edge of the t-cells was found by the edge detection code on this GitHub. Following the curvature calculation, this code can also apply a threshold for the curvature followed by a watershed-based villi detection. Once the villi have been identified, the geometric parameters describing each villus can also be calculated.For this analysis, you will need to have used the previously published edge segmentation code.